members and quantity paramters
